♻️ chore: refactor memoryStore into a helper instead of an IIFE
